requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
mysqlclient==2.2.4
pillow==10.2.0
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
